Move serialization tests to TLE repository test file
